fix(gir): restore compatibility for ints and nullable strings

+++ v4/adw/adw-bottom-sheet.go
@@ -156,13 +156,13 @@ func (x *BottomSheet) GetBottomBar() *gtk.Widget {
 	return cls
 }
 
-var xBottomSheetGetBottomBarHeight func(uintptr) int32
+var xBottomSheetGetBottomBarHeight func(uintptr) int
 
 // Gets the current bottom bar height.
 //
 // It can be used to shift the content upwards permanently to accommodate for
 // the bottom bar.
-func (x *BottomSheet) GetBottomBarHeight() int32 {
+func (x *BottomSheet) GetBottomBarHeight() int {
 	cret := xBottomSheetGetBottomBarHeight(x.GoPointer())
 	return cret
 }
@@ -249,12 +249,12 @@ func (x *BottomSheet) GetSheet() *gtk.Widget {
 	return cls
 }
 
-var xBottomSheetGetSheetHeight func(uintptr) int32
+var xBottomSheetGetSheetHeight func(uintptr) int
 
 // Gets the current bottom sheet height.
 //
 // It can be used to shift the content upwards when the bottom sheet is open.
-func (x *BottomSheet) GetSheetHeight() int32 {
+func (x *BottomSheet) GetSheetHeight() int {
 	cret := xBottomSheetGetSheetHeight(x.GoPointer())
 	return cret
 }
@@ -434,10 +434,10 @@ func (x *BottomSheet) GetPropertyAlign() float32 {
 //
 // It can be used to shift the content upwards permanently to accommodate for
 // the bottom bar.
-func (x *BottomSheet) GetPropertyBottomBarHeight() int32 {
+func (x *BottomSheet) GetPropertyBottomBarHeight() int {
 	var v gobject.Value
 	x.GetProperty("bottom-bar-height", &v)
-	return v.GetLong()
+	return v.GetInt()
 }
 
 // SetPropertyCanClose sets the "can-close" property.
@@ -594,10 +594,10 @@ func (x *BottomSheet) GetPropertyRevealBottomBar() bool {
 // The current bottom sheet height.
 //
 // It can be used to shift the content upwards when the bottom sheet is open.
-func (x *BottomSheet) GetPropertySheetHeight() int32 {
+func (x *BottomSheet) GetPropertySheetHeight() int {
 	var v gobject.Value
 	x.GetProperty("sheet-height", &v)
-	return v.GetLong()
+	return v.GetInt()
 }
 
 // SetPropertyShowDragHandle sets the "show-drag-handle" property.
@@ -635,7 +635,7 @@ func (x *BottomSheet) GetPropertyShowDragHandle() bool {
 
 // Emitted when the close button or shortcut is used while
 // [property@Dialog:can-close] is set to `FALSE`.
-func (x *BottomSheet) ConnectCloseAttempt(cb *func(BottomSheet)) uint32 {
+func (x *BottomSheet) ConnectCloseAttempt(cb *func(BottomSheet)) uint {
 	cbPtr := uintptr(unsafe.Pointer(cb))
 	if cbRefPtr, ok := glib.GetCallback(cbPtr); ok {
 		handlerID := gobject.SignalConnect(x.GoPointer(), "close-attempt", cbRefPtr)
@@ -681,7 +681,7 @@ func (x *BottomSheet) GetProgress() float64 {
 //
 // Each snap point represents a progress value that is considered acceptable to
 // end the swipe on.
-func (x *BottomSheet) GetSnapPoints(NSnapPointsVar *int32) uintptr {
+func (x *BottomSheet) GetSnapPoints(NSnapPointsVar *int) uintptr {
 	cret := XAdwSwipeableGetSnapPoints(x.GoPointer(), NSnapPointsVar)
 	return cret
 }
@@ -765,7 +765,7 @@ func (x *BottomSheet) GetAtContext() *gtk.ATContext {
 // This functionality can be overridden by `GtkAccessible`
 // implementations, e.g. to get the bounds from an ignored
 // child widget.
-func (x *BottomSheet) GetBounds(XVar *int32, YVar *int32, WidthVar *int32, HeightVar *int32) bool {
+func (x *BottomSheet) GetBounds(XVar *int, YVar *int, WidthVar *int, HeightVar *int) bool {
 	cret := gtk.XGtkAccessibleGetBounds(x.GoPointer(), XVar, YVar, WidthVar, HeightVar)
 	return cret
 }
@@ -881,7 +881,7 @@ func (x *BottomSheet) UpdateProperty(FirstPropertyVar gtk.AccessibleProperty, va
 // property change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *BottomSheet) UpdatePropertyValue(NPropertiesVar int32, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
+func (x *BottomSheet) UpdatePropertyValue(NPropertiesVar int, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
 	gtk.XGtkAccessibleUpdatePropertyValue(x.GoPointer(), NPropertiesVar, PropertiesVar, ValuesVar)
 }
 
@@ -913,7 +913,7 @@ func (x *BottomSheet) UpdateRelation(FirstRelationVar gtk.AccessibleRelation, va
 // relation change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *BottomSheet) UpdateRelationValue(NRelationsVar int32, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
+func (x *BottomSheet) UpdateRelationValue(NRelationsVar int, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
 	gtk.XGtkAccessibleUpdateRelationValue(x.GoPointer(), NRelationsVar, RelationsVar, ValuesVar)
 }
 
@@ -946,7 +946,7 @@ func (x *BottomSheet) UpdateState(FirstStateVar gtk.AccessibleState, varArgs ...
 // state change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *BottomSheet) UpdateStateValue(NStatesVar int32, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
+func (x *BottomSheet) UpdateStateValue(NStatesVar int, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
 	gtk.XGtkAccessibleUpdateStateValue(x.GoPointer(), NStatesVar, StatesVar, ValuesVar)
 }
